+++ packages/shorebird_cli/bin/shorebird.dart
@@ -14,7 +14,7 @@ import 'package:shorebird_cli/src/doctor.dart';
 import 'package:shorebird_cli/src/engine_config.dart';
 import 'package:shorebird_cli/src/executables/executables.dart';
 import 'package:shorebird_cli/src/http_client/http_client.dart';
-import 'package:shorebird_cli/src/logger.dart';
+import 'package:shorebird_cli/src/logging/logging.dart';
 import 'package:shorebird_cli/src/network_checker.dart';
 import 'package:shorebird_cli/src/os/os.dart';
 import 'package:shorebird_cli/src/patch_diff_checker.dart';
@@ -34,6 +34,15 @@ Future<void> main(List<String> args) async {
   final loggingStdout = runScoped(
     () => LoggingStdout(baseStdOut: stdout, logFile: currentRunLogFile),
     values: {shorebirdEnvRef},
+  );
+
+  // Write the current command to the top of the log file.
+  currentRunLogFile.writeAsStringSync(
+    '''
+Command: shorebird ${args.join(' ')}
+
+''',
+    mode: FileMode.append,
   );
 
   await IOOverrides.runZoned(