+++ packages/shorebird_cli/lib/src/executables/aot_tools.dart
@@ -9,7 +9,7 @@ import 'package:scoped_deps/scoped_deps.dart';
 import 'package:shorebird_cli/src/cache.dart';
 import 'package:shorebird_cli/src/engine_config.dart';
 import 'package:shorebird_cli/src/extensions/version.dart';
-import 'package:shorebird_cli/src/logger.dart';
+import 'package:shorebird_cli/src/logging/logging.dart';
 import 'package:shorebird_cli/src/shorebird_artifacts.dart';
 import 'package:shorebird_cli/src/shorebird_env.dart';
 import 'package:shorebird_cli/src/shorebird_process.dart';
@@ -243,6 +243,7 @@ class AotTools {
     required String outputPath,
     String? workingDirectory,
     String? dumpDebugInfoPath,
+    List<String> additionalArgs = const [],
   }) async {
     // We use the json lines format. https://jsonlines.org
     const linkJson = 'link.jsonl';
@@ -263,6 +264,10 @@ class AotTools {
           '--redirect-to=${p.join(outputDir, linkJson)}',
         ],
         if (dumpDebugInfoPath != null) '--dump-debug-info=$dumpDebugInfoPath',
+        if (additionalArgs.isNotEmpty) ...[
+          '--',
+          ...additionalArgs,
+        ],
       ],
       workingDirectory: workingDirectory,
     );