+++ packages/shorebird_cli/test/src/shorebird_cli_command_runner_test.dart
@@ -5,7 +5,7 @@ import 'package:mason_logger/mason_logger.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:platform/platform.dart';
 import 'package:scoped_deps/scoped_deps.dart';
-import 'package:shorebird_cli/src/logger.dart' hide logger;
+import 'package:shorebird_cli/src/logging/logging.dart' hide logger;
 import 'package:shorebird_cli/src/platform.dart';
 import 'package:shorebird_cli/src/shorebird_cli_command_runner.dart';
 import 'package:shorebird_cli/src/shorebird_command.dart';
@@ -110,6 +110,20 @@ void main() {
       expect(result, equals(ExitCode.usage.code));
       verify(() => logger.err(exception.message)).called(1);
       verify(() => logger.info(commandRunner.usage)).called(1);
+    });
+
+    group('when runCommand returns null exitCode', () {
+      test('does not print failure text', () async {
+        final result = await runWithOverrides(
+          () => commandRunner.run(['--help']),
+        );
+        expect(result, equals(ExitCode.success.code));
+        verifyNever(
+          () => logger.info(
+            any(that: contains("If you aren't sure why this command failed")),
+          ),
+        );
+      });
     });
 
     test('handles UsageException', () async {